+++ src/main/java/com/abavilla/fpi/entity/impl/gl/GLRewardsResp.java
@@ -18,20 +18,20 @@
 
 package com.abavilla.fpi.entity.impl.gl;
 
+import java.time.LocalDateTime;
+
 import com.abavilla.fpi.entity.mongo.AbsMongoField;
 import io.quarkus.runtime.annotations.RegisterForReflection;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
 @Data
 @EqualsAndHashCode(callSuper = true)
 @RegisterForReflection
 @NoArgsConstructor
 public class GLRewardsResp extends AbsMongoField {
-  private Integer transactionId;
+  private Long transactionId;
   private String status;
   private String address;
   private String promo;
